Use data-duration to initialise duration element
If data-duration is set on the <audio> element, then the default player
interprets this as the content to initialise the duration element of the
player instead of 00:00. If data-duration is an integer, then it is
interpreted as a number of seconds (i.e. 121 becomes 02:01) otherwise
the text is copied verbatim (e.g. allowing --:--).

This is useful on devices which do not preload audio files (e.g. mobile
devices).

The text is overwritten when the file is first played (or when the data
otherwise becomes available).